separate uneficient codes in a file to fix later
